更改build

+++ build.cmd
@@ -5,11 +5,13 @@ if "%ver%"=="" (
     goto getvtag
 )
 if "%1"=="build" (
-    echo [BUILD]building... 1/2
+    echo [BUILD]building... 1/3
     pyinstaller -F -p "." -i dashedgeless.ico -n on_start on_start.py
     copy /Y dist\on_start.exe .\
-    echo [BUILD]building... 2/2
+    echo [BUILD]building... 2/3
     pyinstaller -F -p "." -i dashedgeless.ico -n on_load on_load.py
+    echo [BUILD]building... 3/3
+    pyinstaller -F -p "." -i dashedgeless.ico -n dash dash.py
     copy /Y dist\on_load.exe .\
     rmdir /S /Q build
     rmdir /S /Q dist
@@ -17,17 +19,18 @@ if "%1"=="build" (
 )
 if "%1"=="7z" (
     echo [BUILD]packing... 1/3
-    mkdir "../tmp/dashedgeless"
-    xcopy /I "*" "../tmp/dashedgeless/"
-    copy "dashedgeless.cmd" "../tmp"
-    echo "../dashedgeless_%ver%_dashedgeless (bot).7z">"../tmp/dashedgeless/whitelist.txt"
-    del /Q "../tmp/dashedgeless/*.log"
-    del /Q "../tmp/dashedgeless/.tip*.log"
+    mkdir "tmp/dashedgeless"
+    xcopy /I "*" "tmp/dashedgeless/"
+    copy "dashedgeless.cmd" "tmp"
+    echo "dashedgeless_%ver%_dashedgeless (bot).7z">"tmp/dashedgeless/whitelist.txt"
+    del /Q "tmp/dashedgeless/*.log"
+    del /Q "tmp/dashedgeless/.tip*.log"
     echo [BUILD]packing... 2/3
-    del /Q "../dashedgeless_*_dashedgeless (bot).7z"
-    7z a "../dashedgeless_%ver%_dashedgeless (bot).7z" "../tmp/*"
+    mkdir dist
+    rmdir /S /Q "dist"
+    7z a "dist/dashedgeless_%ver%_dashedgeless (bot).7z" "../tmp/*"
     echo [BUILD]packing... 3/3
-    rmdir /S /Q "../tmp"
+    rmdir /S /Q "tmp"
 )
 if "%1"=="v" (
     echo --- Build tools [dashedgeless] ---
@@ -38,8 +41,4 @@ if "%1"=="" (
     echo version: %ver%
     build build
     build 7z
-)
-if "%1"=="mv" (
-    mkdir dist
-    move /Y "../dashedgeless_%ver%_dashedgeless (bot).7z" "dist/"
 )
